[parsing] camera init done

+++ includes/camera.h
@@ -6,7 +6,7 @@
 /*   By: jwee <jwee@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/13 23:02:54 by jwee              #+#    #+#             */
-/*   Updated: 2023/04/19 23:34:00 by jwee             ###   ########.fr       */
+/*   Updated: 2023/04/24 22:27:08 by jwee             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,11 +18,11 @@
 typedef struct s_cam
 {
 	t_point	origin;
-	t_vec viewpoint;
+	t_vec	viewpoint;
 	double	viewport_height;
 	double	viewport_width;
 	double	focal_length;
-	double	fov;
+	int		fov;
 	t_vec	horizon;
 	t_vec	vertical;
 	t_point	corner;
